chore: took app passcode out of dev experiments
