+++ src/frontend/AppContainerWrapper.tsx
@@ -22,6 +22,9 @@ import {
   persistNavigationState,
 } from "./Navigation/navigationStateHelperFunctions";
 import { AppNavigator } from "./Navigation/AppNavigator";
+import { AppState, AppStateStatus } from "react-native";
+import { useAuthState } from "./hooks/useAuthState";
+import { SecurityContext } from "./context/SecurityContext";
 
 // Turn on logging if in debug mode
 if (__DEV__) debug.enable("*");
@@ -33,6 +36,29 @@ const AppContainerWrapper = () => {
   const [initialNavState, setInitialNavState] = React.useState<
     InitialState | "loading" | undefined
   >("loading");
+
+  const { authValuesSet, setToUnauthenticated, authState } = React.useContext(
+    SecurityContext
+  );
+
+  React.useEffect(() => {
+    const appStateListener = AppState.addEventListener(
+      "change",
+      (nextAppState: AppStateStatus) => {
+        if (authValuesSet.passcodeSet) {
+          if (
+            nextAppState === "active" ||
+            nextAppState === "background" ||
+            nextAppState === "inactive"
+          ) {
+            setToUnauthenticated();
+          }
+        }
+      }
+    );
+
+    return () => appStateListener.remove();
+  }, [authValuesSet.passcodeSet, setToUnauthenticated]);
 
   const navRef = useNavigationContainerRef<AppStackList>();
