+++ src/frontend/context/ObservationsContext.tsx
@@ -3,6 +3,7 @@ import { Observation } from "mapeo-schema";
 
 import api from "../api";
 import { Status } from "../sharedTypes";
+import { useAuthState } from "../hooks/useAuthState";
 
 export interface ObservationAttachment {
   id: string;
@@ -89,6 +90,8 @@ export const ObservationsProvider = ({
     [state]
   );
 
+  const authState = useAuthState();
+
   // This will load observations on first load and reload them every time the
   // value of state.reload changes (dispatch({type: "reload"}) will do this)
   React.useEffect(() => {
@@ -111,6 +114,15 @@ export const ObservationsProvider = ({
       didCancel = true;
     };
   }, [state.reload]);
+
+  React.useEffect(() => {
+    if (authState === "obscured") {
+      dispatch({ type: "reload_success", value: [] });
+      return;
+    }
+
+    dispatch({ type: "reload" });
+  }, [authState]);
 
   return (
     <ObservationsContext.Provider value={contextValue}>
