+++ src/frontend/lib/DevExperiments.ts
@@ -1,6 +1,5 @@
 export const devExperiments = {
   onboarding: process.env.FEATURE_ONBOARDING === "true",
-  appPasscode: process.env.FEATURE_PASSCODE === "true",
   mapSettings: process.env.FEATURE_MAP_SETTINGS === "true",
 } as const;
